requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
idna==2.6
olefile==0.44
Pillow==4.2.1
py==1.4.34
pytest==3.2.2
requests==2.18.4
urllib3==1.22
